Move time.h to lib/extras.
This patch reduces the libjxl_dec.so by about 2 KB thanks to removing
the unconditional call to TraceRunBegin() and TraceRunEnd() when
data parallel tracing is not used, which is the default.

Replaced butteraugli debug usage of Now() with std::chrono::steady_clock
instead. This should be moved to a gbench instead.

Removed the tracing capabilities in data_parallel, this was only enabled
in developer builds when manually modifying the source code to set
DATA_PARALLEL_TRACE to 1.
