+++ lib/extras/time.cc
@@ -3,7 +3,7 @@
 // Use of this source code is governed by a BSD-style
 // license that can be found in the LICENSE file.
 
-#include "lib/jxl/base/time.h"
+#include "lib/extras/time.h"
 
 #include <stdint.h>
 #include <stdio.h>
