+++ lib/lib.gni
@@ -60,8 +60,6 @@ libjxl_dec_sources = [
     "jxl/base/status.cc",
     "jxl/base/status.h",
     "jxl/base/thread_pool_internal.h",
-    "jxl/base/time.cc",
-    "jxl/base/time.h",
     "jxl/blending.cc",
     "jxl/blending.h",
     "jxl/chroma_from_luma.cc",
