+++ tools/benchmark/benchmark_codec_png.cc
@@ -10,11 +10,11 @@
 #include <string>
 
 #include "lib/extras/codec_png.h"
+#include "lib/extras/time.h"
 #include "lib/jxl/base/data_parallel.h"
 #include "lib/jxl/base/padded_bytes.h"
 #include "lib/jxl/base/span.h"
 #include "lib/jxl/base/thread_pool_internal.h"
-#include "lib/jxl/base/time.h"
 #include "lib/jxl/codec_in_out.h"
 #include "lib/jxl/image_bundle.h"
