+++ tools/benchmark/benchmark_codec_webp.cc
@@ -12,12 +12,12 @@
 #include <string>
 #include <vector>
 
+#include "lib/extras/time.h"
 #include "lib/jxl/base/compiler_specific.h"
 #include "lib/jxl/base/data_parallel.h"
 #include "lib/jxl/base/padded_bytes.h"
 #include "lib/jxl/base/span.h"
 #include "lib/jxl/base/thread_pool_internal.h"
-#include "lib/jxl/base/time.h"
 #include "lib/jxl/codec_in_out.h"
 #include "lib/jxl/enc_external_image.h"
 #include "lib/jxl/image.h"
